refactor: User input determines lines of raw data to display
Previously, the number of lines of raw detail records displayed was hard
coded at 5. This was inflexible and did not cater to individual user
preferences. Some users will prefer to see smaller chunks of raw data,
and others will prefer to see larger chunks. This change improves the UI/UX
by asking the user to provide the number of lines they would like to see.